requirements.txt
# Core AI dependencies
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.20.0
pillow>=10.0.0
numpy>=1.24.0

# FastMCP for server-client communication (legacy)
fastmcp>=2.0.0
mcp>=1.0.0

# HTTP server and client
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.28.0

# Web UI
streamlit>=1.28.0

# Development (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0